Deconstructed cart for component re-use outside of drawer
